+++ xently/libraries/data-network/src/main/java/co/ke/xently/libraries/data/network/websocket/StompWebSocketClient.kt
@@ -26,6 +26,7 @@ interface StompWebSocketClient {
     ): Flow<T>
 
     companion object Noop : StompWebSocketClient {
+        const val TAG = "StompWebSocketClient"
         override suspend fun sendMessage(
             url: String,
             submissionDelay: Duration,
